pyproject.toml
[project]
name = "spit-hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "einops>=0.8.2",
    "fastapi>=0.129.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "opencv-contrib-python>=4.13.0.92",
    "opencv-python>=4.13.0.92",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.0",
    "seaborn>=0.13.2",
    "timm>=1.0.24",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.3",
    "ultralytics>=8.4.14",
    "uvicorn[standard]>=0.41.0",
]
